requirements-web.txt
flask==3.0.0
flask-socketio==5.3.5
requests==2.31.0
gdown==5.1.0
beautifulsoup4==4.12.3
face-recognition==1.3.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
python-socketio==5.10.0
eventlet==0.35.1
